+++ AIChat Watch AppTests/AIChat_Watch_AppTests.swift
@@ -12,9 +12,6 @@ import XCTest
 @testable import AIChat_Watch_App
 
 final class AIChat_Watch_AppTests: XCTestCase {
-    @MainActor
-    private static var retainedConversationListPerformanceStores: [ChatStore] = []
-
     func testContextWindowMapsRolesAndPreservesLatestMessages() {
         let imageData = Data([0x01, 0x02, 0x03])
         let attachment = ChatImageAttachment(
@@ -592,67 +589,6 @@ final class AIChat_Watch_AppTests: XCTestCase {
     }
 
     @MainActor
-    func testConversationDetailInitialRenderPerformanceForHeavyCollapsedHistory() async throws {
-        let conversation = ConversationThread(
-            title: "Heavy Markdown",
-            messages: makeHeavyMarkdownMessages(count: 8)
-        )
-        let store = try await makeLoadedStore(conversations: [conversation])
-        let options = XCTMeasureOptions()
-        options.iterationCount = 5
-
-        XCTAssertEqual(
-            ConversationHistoryRenderBudget.visibleMessageCount(
-                in: conversation.messages,
-                budget: 10
-            ),
-            2
-        )
-
-        measure(metrics: [XCTClockMetric()], options: options) {
-            let start = CFAbsoluteTimeGetCurrent()
-            let snapshot = renderConversationDetailSnapshot(
-                store: store,
-                conversationID: conversation.id
-            )
-            let elapsed = CFAbsoluteTimeGetCurrent() - start
-
-            XCTAssertNotNil(snapshot)
-            XCTAssertLessThan(
-                elapsed,
-                0.35,
-                "Heavy conversation initial render regressed to \(elapsed)s"
-            )
-        }
-    }
-
-    @MainActor
-    func testConversationDetailInitialRenderPerformanceForHeavyLatexHistory() async throws {
-        let conversation = ConversationThread(
-            title: "Heavy LaTeX",
-            messages: makeHeavyLatexMessages(count: 6)
-        )
-        let store = try await makeLoadedStore(conversations: [conversation])
-        let options = XCTMeasureOptions()
-        options.iterationCount = 5
-
-        measure(metrics: [XCTClockMetric()], options: options) {
-            let start = CFAbsoluteTimeGetCurrent()
-            let snapshot = renderConversationDetailSnapshot(
-                store: store,
-                conversationID: conversation.id
-            )
-            let elapsed = CFAbsoluteTimeGetCurrent() - start
-
-            XCTAssertNotNil(snapshot)
-            XCTAssertLessThan(
-                elapsed,
-                0.35,
-                "Heavy LaTeX conversation initial render regressed to \(elapsed)s"
-            )
-        }
-    }
-
     func testConversationThreadListSummaryCapturesPreviewAndAttachmentFlags() {
         let conversation = ConversationThread(
             title: "Summary",
@@ -692,61 +628,6 @@ final class AIChat_Watch_AppTests: XCTestCase {
         XCTAssertEqual(summary.messageCount, 3)
         XCTAssertTrue(summary.containsAudioAttachments)
         XCTAssertTrue(summary.containsImageAttachments)
-    }
-
-    @MainActor
-    func testConversationListInitialRenderPerformanceForLargeHistory() async throws {
-        let options = XCTMeasureOptions()
-        options.iterationCount = 5
-
-        measure(metrics: [XCTClockMetric()], options: options) {
-            let start = CFAbsoluteTimeGetCurrent()
-            let snapshot = autoreleasepool { () -> CGImage? in
-                let store = ChatStore.previewStore(
-                    conversations: makeHistoryListConversations(count: 180),
-                    completionFeedbackProvider: NoopCompletionFeedbackProvider()
-                )
-                Self.retainedConversationListPerformanceStores.append(store)
-                return renderConversationListSnapshot(store: store)
-            }
-            let elapsed = CFAbsoluteTimeGetCurrent() - start
-
-            XCTAssertNotNil(snapshot)
-            XCTAssertLessThan(
-                elapsed,
-                0.45,
-                "Large watch conversation list initial render regressed to \(elapsed)s"
-            )
-        }
-    }
-
-    @MainActor
-    func testConversationListContinuationRenderPerformance() async throws {
-        let options = XCTMeasureOptions()
-        options.iterationCount = 5
-
-        measure(metrics: [XCTClockMetric()], options: options) {
-            let start = CFAbsoluteTimeGetCurrent()
-            let snapshot = autoreleasepool { () -> CGImage? in
-                let store = ChatStore.previewStore(
-                    conversations: makeHistoryListConversations(count: 180),
-                    completionFeedbackProvider: NoopCompletionFeedbackProvider()
-                )
-                Self.retainedConversationListPerformanceStores.append(store)
-                return renderConversationListSnapshot(
-                    store: store,
-                    initialVisibleConversationLimit: 12
-                )
-            }
-            let elapsed = CFAbsoluteTimeGetCurrent() - start
-
-            XCTAssertNotNil(snapshot)
-            XCTAssertLessThan(
-                elapsed,
-                0.28,
-                "Watch conversation list continuation render regressed to \(elapsed)s"
-            )
-        }
     }
 
     @MainActor
@@ -2926,157 +2807,6 @@ final class AIChat_Watch_AppTests: XCTestCase {
         )
         XCTAssertEqual(persistedConversation.messages.count, 1)
         XCTAssertEqual(persistedConversation.messages.first?.role, .user)
-    }
-
-    @MainActor
-    private func makeLoadedStore(
-        conversations: [ConversationThread]
-    ) async throws -> ChatStore {
-        let configuration = AppConfiguration(
-            backendMode: .direct,
-            geminiAPIKey: "test",
-            geminiModel: "gemini-3.1-pro-preview",
-            geminiTranscriptionModel: "gemini-3-flash-preview",
-            relayBaseURL: nil,
-            relayBearerToken: nil,
-            relayStreamPath: "v1/chat/stream",
-            appGroupIdentifier: nil
-        )
-        let repository = ConversationRepository(
-            configuration: configuration,
-            rootURL: FileManager.default.temporaryDirectory
-                .appendingPathComponent("AIChatPerf-\(UUID().uuidString)", isDirectory: true)
-        )
-
-        for conversation in conversations {
-            try await repository.save(conversation)
-        }
-
-        let store = ChatStore(
-            repository: repository,
-            aiService: EchoReplyAIStreamingService(),
-            transcriptionService: nil,
-            configuration: configuration,
-            syncBridge: CompanionSyncBridge()
-        )
-        await store.loadConversationsIfNeeded()
-        return store
-    }
-
-    @MainActor
-    private func renderConversationDetailSnapshot(
-        store: ChatStore,
-        conversationID: UUID
-    ) -> CGImage? {
-        let width: CGFloat = 184
-        let height: CGFloat = 224
-        let content = ConversationDetailView(conversationID: conversationID)
-            .environmentObject(store)
-            .frame(width: width, height: height)
-
-        let renderer = ImageRenderer(content: content)
-        renderer.scale = 2
-        renderer.proposedSize = ProposedViewSize(width: width, height: height)
-        return renderer.cgImage
-    }
-
-    @MainActor
-    private func renderConversationListSnapshot(
-        store: ChatStore,
-        initialVisibleConversationLimit: Int = 0
-    ) -> CGImage? {
-        let width: CGFloat = 184
-        let height: CGFloat = 224
-        let content = ConversationListView(
-            navigationPath: .constant([]),
-            initialVisibleConversationLimit: initialVisibleConversationLimit
-        )
-            .environmentObject(store)
-            .frame(width: width, height: height)
-
-        let renderer = ImageRenderer(content: content)
-        renderer.scale = 2
-        renderer.proposedSize = ProposedViewSize(width: width, height: height)
-        return renderer.cgImage
-    }
-
-    private func makeHistoryListConversations(count: Int) -> [ConversationThread] {
-        let seededDate = Date(timeIntervalSince1970: 1_762_401_000)
-
-        return (0..<count).map { index in
-            let createdAt = seededDate.addingTimeInterval(Double(index * 90))
-            let title = "History Thread \(index + 1)"
-            return ConversationThread(
-                id: UUID(),
-                title: title,
-                createdAt: createdAt,
-                updatedAt: createdAt.addingTimeInterval(30),
-                isFavorite: index.isMultiple(of: 11),
-                messages: [
-                    ChatMessage(
-                        role: .assistant,
-                        text: """
-                        \(title)
-                        This seeded history row includes enough text to keep watch list cells realistic during render testing.
-                        """,
-                        createdAt: createdAt
-                    )
-                ]
-            )
-        }
-        .sorted(by: ConversationThread.sortsByMostRecentFirst)
-    }
-
-    private func makeHeavyMarkdownMessages(count: Int) -> [ChatMessage] {
-        let heavyMarkdown = Array(
-            repeating: """
-            ## 推导步骤
-            - 先整理条件 `value`
-            - 再计算公式 $$\\int_0^1 x^2 dx = \\frac{1}{3}$$
-
-            ```swift
-            let answer = 42
-            print(answer)
-            ```
-
-            这一段用于模拟真实的大段 markdown 与 LaTeX 回复内容。
-            """,
-            count: 36
-        ).joined(separator: "\n\n")
-
-        return (0..<count).map { index in
-            ChatMessage(
-                role: index.isMultiple(of: 2) ? .user : .assistant,
-                text: index.isMultiple(of: 2) ? "继续第 \(index + 1) 步" : heavyMarkdown
-            )
-        }
-    }
-
-    private func makeHeavyLatexMessages(count: Int) -> [ChatMessage] {
-        let heavyLatex = Array(
-            repeating: """
-            ## 推导
-            令 $a_n = \\frac{1}{n^2 + 1}$，并比较下式：
-
-            $$
-            f(x) = \\sum_{k=1}^{18} \\frac{x^k}{k!}
-            $$
-
-            \\[
-            \\int_0^1 \\frac{1}{1 + x^2} dx = \\frac{\\pi}{4}
-            \\]
-
-            再验证 $\\alpha^2 + \\beta^2 = \\gamma^2$ 的近似边界。
-            """,
-            count: 8
-        ).joined(separator: "\n\n")
-
-        return (0..<count).map { index in
-            ChatMessage(
-                role: index.isMultiple(of: 2) ? .user : .assistant,
-                text: index.isMultiple(of: 2) ? "继续公式第 \(index + 1) 步" : heavyLatex
-            )
-        }
     }
 
     @MainActor
